Implement persistence layer

+++ src/main/java/ohnana/model/generic/AttributeInterface.java
@@ -1,5 +1,5 @@
 package ohnana.model.generic;
 
 public interface AttributeInterface<T> {
-    public int getId();
+    public Long getId();
 }
